requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.8
python-magic==0.4.27
redis==5.0.1
pymongo==4.6.1
dnspython==2.4.2
gunicorn==21.2.0
python-multipart==0.0.6
